backend/requirements.txt
chromadb==1.0.9
email-validator>=2.0.0
flask>=2.0.0
flask-sqlalchemy>=3.0.0
groq>=0.24.0
gunicorn>=23.0.0
pillow>=11.0.0
psycopg2-binary>=2.9.0
pymupdf>=1.25.0
pypdf2>=3.0.0
python-dotenv>=1.0.0
werkzeug>=2.0.0